virida_pricing_api_service/requirements.txt
fastapi==0.61.1
uvicorn
pydantic
sqlalchemy
pymysql
cryptography
pyjwt
alembic
requests
#numpy
#scipy
#tensorflow
aiohttp
pytest
pytest-asyncio
paramiko
#pandas==1.1.4
fredapi==0.4.3
google-cloud-error-reporting
sendgrid~=6.6.0
matplotlib
